Added a interpreter for hH and a type checker.
The example prv.def describes a small interpreter for hereditary
Harrop formulas.  

The file pi.def has been modified to include a type checker for
pi-calculus expressions.  It seemed that a good way to do this was to
use the lambda Prolog implementation of simple type checking and then
use the interpreter in prv.def to execute the type checker.  I also
reorganized the pi.def file a bit.  I did not try out the weak
bisimulation stuff yet.

I added the lists.def file with the goal of eventually having it store
most of the basic list manipulation functions that we need.