Some small changes to various userguide and annoucement (added the concurrency mailing list on the recommendation of a model checking colleague).
Rewrote the bedwyr homepage and associated .css file.  Leave the rest
of the homepage to David.  Links to the document and tarbal need to be
added.  Feel free to make other changes as you see fit.  I have put
the home page into the official htdocs location so they are no on the
web yet.