1) I introduce two variants of conjunction. These are *not* new connectives; logically they are still the good-old conjunction. But they allow me to use different search strategies to prove conjunctive queries. These are mainly used to detect finite failure more cleverly.
 (a) Independent finite-failure: Syntax A ^ B. The operational meaning 
of this is as follows: first try proving A and B independently. If one 
of the finitely fails, then the original goal is not provable. 
Otherwise,  restart with the new goal A & B.

 (b) Finite-failure first: Syntax (A | B).  Its meaning is as follows: 
first try prove A and B successively. If B fails, then restart with the 
new goal (B & A). More generally, if the goal is
(A1 | (A2 | (A3 | (... | An))))
then it tries proving successively Ai, and stop at the point where Ai 
fails (e.g., i = 3), and restart with  
(A3 | A1 | .....)

2) Added a couple more examples to show how these new operators are 
used. Moved these examples to a new directory examples/constraints.