found a magic bug™
Among the modifications I made when I introduced pre-terms,
I changed the order in which I add constraints to a clause
(p X X (f Y) := body  -->  p X U V := U = X /\ V = (f Y) /\ body),
which shouldn't have the slightest consequence on the computations
(except for alpha-conversion and what usually happens with lazy conjunction).
But then the fourth assertion of lambda.def failed,
and it took me a week to find out that this insignificant change was the culprit.
It is possible to see this bug before my work, at rev 703, by changing
"wt A B C" by "wt A C B". I can't for the life of me understand how this
can be, so I promoted this bug to the rank of magic one™, and taking
care of it one day™ was my first good resolution.