no higher order in quantifiers
Like announced in commit #738, the type of free variables and variables
bound by quantifiers cannot contain TProp.
This is an improvement on commit #722, which only dealt with free
variables which type was exactly TProp:
- "Cannot give bound variable %s the type prop" was replaced by
  "Typing error: cannot quantify over type %a"
- the check is done when the context of the variable has been
  type-checked instead of when it is seen for the first time:
  * after the type checking/inference of the quantified formula
  * after the whole checking for a free variable

This means the error on a type will be hidden by the first of any other
errors appearing on the context, even if this error comes after:
- "exists (X : prop), p X." and "exists X, X /\ p X." both raise
  "Undeclared constant or predicate p."
- "exists (X : prop), true." and "exists X, X." both raise
  "Typing error: cannot quantify over type prop."
- "X." raises "Typing error: cannot give free variable X the type prop"

Theoretically, this shouldn't forbid higher order arguments for
predicates (or even constants, for that matter), but right now, during
the definition of p, the clause "p X := X" has one forbidden free variable
in its head and doesn't type-check.