v1.3-beta1: first batch of updates examples
With some new kind and type declarations, a few ugly '$' added in front
of some names conflicting with keywords ("env", "type", etc) and some
higher-order arguments to avoid the need for incremental definitions
(cf the first argument of "one" and "onep" in pi/pi.def, for instance),
they seem to work.
Of course, more assertions would be needed to check that the general
behaviour is really the same.