free propositional variables allowed with #typeof
A small amend to commits 743/746: the checks cannot raise
"cannot give free variable <<blah>> the type <<... -> prop -> ...>>"
if the checked term is temporary (ie if the results of the type
inference will be discarded).
Currently, this means that "#typeof X /\ Y." will give the correct type
for X and Y, even if the same term cannot be queried on.
This doesn't change what would happen if there was a quantification over
the problematic variables (still "cannot quantify over...").