infix constants are right-associative
Unfortunately, a bug in ocamlyacc (which has been unmaintained for
years, btw) keeps us from throwing syntax errors when a non-associative
operator is not treated as such (by default they are *always*
left-associative, even if the doc claims that they are not if
precedences aren't specified), so since we have to use associativity,
consistency with "->" will be fine.

Also, I merged term_atoms and form_atoms, since the distinction seemed
unnecessary, and it shouldn't change anything from the outside.