corrected the magic bug™ (cf rev 740)
A fix on timestamps, implemented in rev 176 and slightly changed in rev
217, was buggy.
Say we are checking A -> B with a timestamp n. We found some solutions
for A, say s1 (timestamped n1) and s2 (n2). We are now trying to prove
B/s1 (n1) and B/s2 (n2). We have to use n1 and n2, and not n (which is
smaller), in order to account for the existential quantifiers we found
inside of A while computing s1 and s2 (basically, we do as if we took
out of B all the quantifiers on which the substitution depend, including
(n1-n), resp. (n2-n), existential quantifiers which become universal
ones).
This is what was done, but with a small bug in the computation of the
timestamp of a solution of a conjunction (and similarly implication,
since we prove B/s1 /\ B/s2 /\ ..., but only the case of conjunction
raised the problem): when checking A1 /\ A2, once we have found that s1
(n1) is a solution for A1, we refine it by searching for s2 (n2),
solution for A2,
*and the timestamp of the global solution is max(n1,n2)* and not just n2
as it was done. This is why swapping two conjuncts could "hide" a n1>n2,
and bring us to try to prove B by depending on eigen variables with
timestamp n1 while being in (and declaring logic variables of) timestamp
only n2, thus raising erroneous OccursChecks.