"#reset." also removes constants' types and types' kinds
Terms and types are not removed but unreferenced. Terms should be GC-ed
(but I'm not sure this is bug-free, Term.Hint might prevent that), but
the type environment is known never to be cleansed. Since old types
won't be reused, this is a regular memory leak, not a #reset problem,
and it is on my TODO list.