change in Table interface (s/find+add/access/)
The change in Table reflects a refactorization of Index code.
Mixing find and add helps for instance to unify the handling of
variables (there was no "when allow_eigenvar" in find):

Define inductive p : nat -> prop by p X.
?= forall x, p x.
?= forall x, (p x -> true).

The first query tables the atom "forall x, p x",
but the second shouldn't use this atom (it could use "exists x, p x" if
we were to table it, though). This inconsistency was corrected.

However, all this is still unclear and incomplete (and, for what it's
worth, incorrect), so more on this later.