added support for exit statuses
There is now a "bad input" state which is entered on some errors
(all that's a matter of syntax and namespace). This state prevents
bedwyr to solve queries and test assertions, but allows further parsing,
so that all the typos of a file may be displayed at once.
Furthermore, bedwyr will not enter the interactive toplevel (though it
will parse the non-interactive toplevel, ie the "-e 'form.'"
command-line options), instead it will exit with a status 2.

There are still some errors that are too bad for bedwyr to keep going
(Llambda problems, failed assertions), in which case the exit status is
3.

Last, some errors aren't sorted (this will include rebel "assert false"
and other bugs) and output more general error messages and exit with
status 1.

In order to keep the old die-already behaviour, a strict mode can be set
at compile time, in which case bedwyr quits immediately with status 2
instead of entering the "bad input" state. Soon to come, a command-line
flag to set this strict mode.
A new example file, examples/internal/input_errors.def, show a list of
"soft" error messages, and a last one that is "hard" and forces exit.