v1.4-alpha3 (Windows installer generation automated)
A real installer with registry keys, uninstaller and start-menu entries.
It is not quite useful yet, because the Windows console is dreadful, and
the look-and-feel is a little old, but still it seems nice to have.

The new file versioninfo.rc doesn't work (but it's only supposed to add
some meta-information to the Windows binary, so nobody cares).