'#include' correction
The list of included files didn't account for changes of the current
directory. So if "test.def" was already included, and another file in
another directory wanted to include another "test.def", it was silently
ignored. It is now verbosely included, and real obvious duplicates are
verbosely skipped.

This is still broken (non-obvious duplicates are not skipped), and will
stay as such: it is not possible to find in a simple and portable (and
compete) way whether two paths denote a same file. This is a small
problem under Linux, where "test.def", "./test.def" and ".//test.def"
are different paths to the same file. It is a bigger problem under
Windows, where both slashes and backslahes are used. The sound and
unsafe approach used here is to try to load files if they seem to be
different, at the risk of provoking a "type already declared" or a "name
conflict".

By the way, the filename is added to the list before the include is
done, as it was done before I mis-corrected it, so no more "Too many
open files" with inclusion loops.