pseudo syntax support for abella proofs in definition files
All that follows a Theorem command is ignored
(no typing, no syntactic checks, barely enough lexical checks to detect
a single Qued command).
A misplaced Qed (with no previous Theorem, or with another command in
between) raises an error.

The vim syntax highlighting is quirky (the first tactic of a line is
colored as a theorem something was colored as a theorem at the end of
the previous line, ie the first tactic of the proof is usually not
commented out) on some vim version. In addition to that, the proof must
not be empty; this will do for now.