[alwen/rebased from poly] I have decided to restrict the polymorphic typing in definitions. Now I forbid a defined predicate to be used polymorphically in its definition.
The previous version was unsound because unification does not preserve
types, and could result in ill-typed terms in answer substitutions.

The question is whether we want a full "let-polymorphism" in definition
clauses (where a defined predicate can be used polymorphically in its
definition). I think to do this we might need to attach types to variables
and do type checking in unification. There could be potentially significant
performance penalty if we do that.

Anyway, I am not even completely sure if the current restriction
preserves typing during runtime. It seems okay on a few examples.