[rebased from poly] improvements of the polymorphism
I. code
- merged TConst and TFunc (both are constant type constructors,
  with the first having no arity)
- split TParam in TParam and TVar (at least)
- parser modified to remove conflicts

II. polymorphism vs inference
The first approach (TParam used for all variablish types, TVar unused)
was in conflict with the fact that some types are temporarily left
unspecified (types of quantified variables, sometimes types of constants
or predicates if the placeholder _ is used) but are not to be considered
as polymorphic.
A good example of this problem is the query
  ?= forall x, x = 42 /\ x = "hello world".
which should raise a typing error but didn't in the last commit, since x
was assumed to be polymorphic and its type was instantiated to two
different fresh types.
On the other hand, using only one type instantiation for x (or a
predicate) for a whole query/formula/clause meant that a predicate
couldn't be fully used polymorphically in its definition (cf the
comments in examples/poly/gadt.def in the previous commit).
A solution is to use TParam for inference and TVar for polymorphism, and
to instantiate the set of TVar's of the type of an object with a new set
of TParam's each time the object is used. This has been implemented with
the existing TParam (displayed "?...") and TVar (displayed "'...").
- "#env." shows the raw types, for instance
  "::: : '1 -> (list '1) -> list '1" for cons;
  a constant declared by "Type c _ -> _." whould show "c : ?42 -> ?43"
  until a query infers the TParam's ?42 and ?43
- "#typeof (:::)." shows the types instantiated (since #typeof can be
  used with a formula containing any number of instances of the
  polymorphic object), for instance
    ?= #typeof (:::).
    "::: : ?42 -> (list ?42) -> list ?42"
    ?= #typeof X = (1 ::: nil) /\ Y = ("1" ::: nil).
    X = (::: 1 nil) /\ Y = (::: "1" nil) : prop
       X : list nat
       Y : list string
In OCaml words, ?42 is '_a and '42 is 'a.
All this works on the original gadt.def example.

III. misc
Ground-checking was corrected and re-enabled. Since it checks that no
TParam is present and TVar's are used for polymorphism, it doesn't fail
on polymorphic objects anymore.
Its aim is to forbid definitions such as
  Define p by p X X.                    % seams polymorphic but isn't
while allowing
  Define p by p.                        % p := true
or
  Define p : _ -> _ by p "vrai".        % p : string -> prop