better consistency in polymorphic type annotations (cf teyjus bug #70)
The formula "forall (x : A) (y : A), x y = Z" was well-typed, since the
scope of a type parameters was the type annotation and not the formula
(ie this was read as "forall (x : A) (y : B), x y = Z"). It know raises
a typing error due to an occurs-check violation.

This is not a bug per se (no-one never said anything about implicit type
quantifiers moving outside of the type to the begining of the formula),
just a non-intuitive behaviour (it was just reported as a Teyjus bug).

Since a fix was added 7 commits ago to make the display of types
formula-scoped in #typeof, this fix makes perfect sense too.