new meta-command: #export
  ?= pred X.
  ?= #export "file.xml"

This will put all atoms tabled by previous calls in an XML file.

Bear in mind that although "#clear_tables" is harmless (you will just
end up with an as-good-as-empty file), "#clear_table pred" removes all
pred-atoms but leave all other atoms, some of which might be linked to
some defunct pred-atoms.
This obviously leads to a runtime crash, but shouldn't be a problem for
your average user ("#clear_table pred" is not your average command).