add contrib files for the XML output (DTD/XSL/CSS)
respectively for validating, transforming and displaying

NB: most browser (ie even IE, but not Konqueror) can render an XML as
HTML if the XSL is present, but Chrome (at least) won't do it if both
files are on the local computer (for security reasons).
The workaround is to run
  `xsltproc --path contrib/xml my-file.xml > my-file.html`
to create the html locally (and to add the CSS next to it).